+++ Solomon_Server/Bulletin_Server/Common/ComDef.cs
