+++ Solomon_Server/Bulletin_Server/DataBase/DBManager.cs
