+++ Solomon_Server/Bulletin_Server/DataBase/ModelBase.cs
