+++ Solomon_Server/Bulletin_Server/JWT/Models/IAuthContainerModel.cs
