+++ Solomon_Server/Bulletin_Server/Models/Meal/Result.cs
