+++ Solomon_Server/Bulletin_Server/Models/Meal/Row.cs
