+++ Solomon_Server/Bulletin_Server/Solomon_Server.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{677C1E29-9CD5-455A-B0B4-C4932862FC3E}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>Bulletin_Server</RootNamespace>
-    <AssemblyName>Bulletin_Server</AssemblyName>
+    <RootNamespace>Solomon_Server</RootNamespace>
+    <AssemblyName>Solomon_Server</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -145,6 +145,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Common\ComDef.cs" />
+    <Compile Include="Common\ResponseMessage.cs" />
     <Compile Include="DataBase\DBManager.cs" />
     <Compile Include="DataBase\ModelBase.cs" />
     <Compile Include="Interface\IService.cs" />
