+++ Solomon_Server/Solomon_Server.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30413.136
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bulletin_Server", "Bulletin_Server\Bulletin_Server.csproj", "{677C1E29-9CD5-455A-B0B4-C4932862FC3E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Solomon_Server", "Bulletin_Server\Solomon_Server.csproj", "{677C1E29-9CD5-455A-B0B4-C4932862FC3E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
